+++ src/main/java/org/generals/domain/BuskingVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class BuskingVO {
 	
-	private Long bkno, rno, bno;
+	private Integer bkno, rno, bno;
 	private BoardVO boardVO;
 	private ReserveVO reserveVO;
 }
